+++ app/src/main/java/com/kau/ttokttok/_core/domain/DomainModule.kt
@@ -1,10 +1,11 @@
 package com.kau.ttokttok._core.domain
 
 import com.kau.ttokttok.data.local.repository.CommunityRepositoryImpl
-import com.kau.ttokttok.domain.repository.CommunityRepository
+import com.kau.ttokttok.data.local.repository.ReportRepositoryImpl
 import com.kau.ttokttok.domain.usecase.community.CreatePostUseCase
 import com.kau.ttokttok.domain.usecase.community.GetPostDetailUseCase
 import com.kau.ttokttok.domain.usecase.community.LoadPostsUseCase
+import com.kau.ttokttok.domain.usecase.monthreport.GetMonthReportUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -13,7 +14,6 @@ import dagger.hilt.components.SingletonComponent
 @Module
 @InstallIn(SingletonComponent::class)
 class DomainModule {
-
 
     // 공지 사항
     @Provides
@@ -35,5 +35,13 @@ class DomainModule {
         repository: CommunityRepositoryImpl
     ): LoadPostsUseCase {
         return LoadPostsUseCase(repository)
+    }
+
+    // 월간 리포트
+    @Provides
+    fun provideGetMonthReportUseCase(
+        repository: ReportRepositoryImpl
+    ): GetMonthReportUseCase {
+        return GetMonthReportUseCase(repository)
     }
 }
