+++ app/src/main/java/com/kau/ttokttok/data/local/repository/ReportRepositoryImpl.kt
@@ -3,8 +3,8 @@ package com.kau.ttokttok.data.local.repository
 import com.kau.ttokttok._core.network.result.NetworkResult
 import com.kau.ttokttok._core.network.result.safeApiCall
 import com.kau.ttokttok.data.remote.api.ReportApiService
+import com.kau.ttokttok.data.remote.dto.report.res.GetApartMonthReportRes
 import com.kau.ttokttok.data.remote.dto.report.res.GetApartmentStatsRes
-import com.kau.ttokttok.data.remote.dto.report.res.GetMonthlyReportRes
 import com.kau.ttokttok.domain.repository.ReportRepository
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -16,6 +16,15 @@ class ReportRepositoryImpl @Inject constructor(
     override suspend fun getApartmentStats(keyword: String): NetworkResult<GetApartmentStatsRes> =
         safeApiCall { api.getApartmentStats(keyword) }
 
-    override suspend fun getMonthlyReport(date: String): NetworkResult<GetMonthlyReportRes> =
-        safeApiCall { api.getMonthlyReport(date)}
+    override suspend fun getMonthlyReport(date: String): GetApartMonthReportRes {
+        return when (val result = safeApiCall { api.getMonthlyReport(date) }) {
+            is NetworkResult.Success -> {
+                result.data
+            }
+
+            is NetworkResult.Error -> {
+                throw Exception(result.message ?: "Network Error")
+            }
+        }
+    }
 }
