+++ app/src/main/java/com/kau/ttokttok/data/remote/api/ReportApiService.kt
@@ -2,8 +2,8 @@ package com.kau.ttokttok.data.remote.api
 
 import com.kau.ttokttok._core.network.model.ApiResponse
 import com.kau.ttokttok.data.remote.dto.report.res.CreateReportRes
+import com.kau.ttokttok.data.remote.dto.report.res.GetApartMonthReportRes
 import com.kau.ttokttok.data.remote.dto.report.res.GetApartmentStatsRes
-import com.kau.ttokttok.data.remote.dto.report.res.GetMonthlyReportRes
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
@@ -18,7 +18,7 @@ interface ReportApiService {
     @GET("api/reports/monthly")
     suspend fun getMonthlyReport(
         @Query("date") date: String
-    ): ApiResponse<GetMonthlyReportRes>
+    ): ApiResponse<GetApartMonthReportRes>
 
     // 리포트 생성(전송)
     @POST("noise/records/{recordId}/send")
