+++ app/src/main/java/com/kau/ttokttok/domain/repository/ReportRepository.kt
@@ -1,10 +1,10 @@
 package com.kau.ttokttok.domain.repository
 
 import com.kau.ttokttok._core.network.result.NetworkResult
+import com.kau.ttokttok.data.remote.dto.report.res.GetApartMonthReportRes
 import com.kau.ttokttok.data.remote.dto.report.res.GetApartmentStatsRes
-import com.kau.ttokttok.data.remote.dto.report.res.GetMonthlyReportRes
 
 interface ReportRepository {
     suspend fun getApartmentStats(keyword: String): NetworkResult<GetApartmentStatsRes>
-    suspend fun getMonthlyReport(date: String): NetworkResult<GetMonthlyReportRes>
+    suspend fun getMonthlyReport(date: String): GetApartMonthReportRes
 }
