+++ app/src/main/java/com/kau/ttokttok/domain/usecase/monthreport/GetMonthReportUseCase.kt
@@ -1,0 +1,18 @@
+package com.kau.ttokttok.domain.usecase.monthreport
+
+import com.kau.ttokttok.domain.repository.ReportRepository
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class GetMonthReportUseCase @Inject constructor(
+    private val repository: ReportRepository
+) {
+    suspend operator fun invoke() = runCatching {
+        val now = LocalDate.now().minusMonths(1).format(DateTimeFormatter.ISO_DATE)
+
+        repository.getMonthlyReport(now)
+    }
+}
