+++ app/src/main/java/com/kau/ttokttok/ui/component/monthreport/NoiseSummaryCard.kt
@@ -1,0 +1,96 @@
+package com.kau.ttokttok.ui.component.monthreport
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.TrendingDown
+import androidx.compose.material.icons.automirrored.filled.TrendingUp
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.kau.ttokttok.ui.theme.*
+
+@Composable
+fun NoiseSummaryCard(
+    modifier: Modifier = Modifier,
+    totalReports: Int = 47,
+    comparedToPrevious: Int = -15, // +면 증가, -면 감소
+) {
+    val isIncrease = comparedToPrevious > 0
+    val trendColor = if (isIncrease) Red400 else Green400
+    val trendIcon = if (isIncrease) Icons.AutoMirrored.Filled.TrendingUp else Icons.AutoMirrored.Filled.TrendingDown
+    val changeLabel = if (isIncrease) "증가" else "감소"
+
+    Card(
+        modifier = modifier
+            .fillMaxWidth(),
+        colors = CardDefaults.cardColors(containerColor = Slate800),
+        shape = RoundedCornerShape(16.dp),
+        border = BorderStroke(0.5.dp, Slate700),
+        elevation = CardDefaults.cardElevation(0.dp)
+    ) {
+        Column(
+            modifier = Modifier.padding(24.dp)
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = 16.dp),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Column {
+                    Text(
+                        text = "이달의 소음 리포트",
+                        color = Gray400,
+                        fontSize = 13.sp,
+                        modifier = Modifier.padding(bottom = 4.dp)
+                    )
+                    Text(
+                        text = "${totalReports}건",
+                        color = Gray100,
+                        fontSize = 28.sp,
+                        fontWeight = FontWeight.Bold
+                    )
+                }
+
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(6.dp)
+                ) {
+                    Icon(
+                        imageVector = trendIcon,
+                        contentDescription = if (isIncrease) "증가" else "감소",
+                        tint = trendColor,
+                        modifier = Modifier.size(28.dp)
+                    )
+                    Text(
+                        text = "${kotlin.math.abs(comparedToPrevious)}%",
+                        color = trendColor,
+                        fontSize = 22.sp,
+                        fontWeight = FontWeight.Medium
+                    )
+                }
+            }
+
+            Text(
+                text = "전월 대비 $changeLabel",
+                color = Gray400,
+                fontSize = 13.sp
+            )
+        }
+    }
+}
