+++ app/src/main/java/com/kau/ttokttok/ui/theme/Color.kt
@@ -1,0 +1,38 @@
+package com.kau.ttokttok.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+// ---------------------
+// Gray Scale
+// ---------------------
+val Gray100 = Color(0xFFF3F4F6)
+val Gray300 = Color(0xFFD1D5DB)
+val Gray400 = Color(0xFF9CA3AF)
+
+// ---------------------
+// Slate Scale
+// ---------------------
+val Slate900 = Color(0xFF0F172A)
+val Slate800 = Color(0xFF1E293B)
+val Slate700 = Color(0xFF334155)
+
+// ---------------------
+// Blue Scale
+// ---------------------
+val Blue400 = Color(0xFF60A5FA)
+val Blue500 = Color(0xFF3B82F6)
+val Blue600 = Color(0xFF2563EB)
+val Blue900 = Color(0xFF1E3A8A)
+
+// ---------------------
+// Purple Scale
+// ---------------------
+val Purple300 = Color(0xFFD8B4FE)    // 밝음
+val Purple700 = Color(0xFF6D28D9)
+val Purple900 = Color(0xFF4C1D95)     // 진함
+
+// ---------------------
+// Status
+// ---------------------
+val Red400 = Color(0xFFF87171)
+val Green400 = Color(0xFF34D399)
